requirements.txt
async-timeout==4.0.3
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
ecdsa==0.18.0
Flask==3.0.0
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
oauthlib==3.2.2
packaging==23.2
pyasn1==0.5.1
python-jose==3.3.0
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.2.1
six==1.16.0
soupsieve==2.5
tenacity==8.2.3
urllib3==2.1.0
Werkzeug==3.0.1
